backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
tensorflow-cpu==2.15.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
scikit-learn==1.3.0
pyjwt==2.8.0
eventlet==0.33.3